==> code/archived/14_fb_impute_18_from_20.py <==
import sklearn.model_selection as ms
from sklearn.pipeline import Pipeline
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.calibration import CalibratedClassifierCV
from sklearn.naive_bayes import MultinomialNB
from sklearn import metrics
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
import pandas as pd
import numpy as np
from joblib import dump, load

# the data required for this processes this script is used for
# requires television data which we are contractually unable to 
# share through Github. However, you can request this data by following the 
# instructions at the following link!
# https://mediaproject.wesleyan.edu/dataaccess/

df = pd.read_csv("data/fb_18_20_for_imputation.csv")

main_cols = ["alt", "transcript"]
issue_cols = [c for c in df.columns if 'ISSUE' in c]

# Loop over all issues
for g in issue_cols:

  # Subset to only the relevant column
  traintestinf = df[main_cols + [g]]
  # Separate all the rows without NAs
  # and make them the train/test set
  traintest = traintestinf.dropna(axis = 0)
  # The rows with NAs are the inference set
  inference = traintestinf[traintestinf[g].isna()]
  
  # If the inference set contains more than 0 rows
  if len(inference) > 0:
    
      # Split
      train, test = ms.train_test_split(traintest, train_size = 0.9)
      # Train & test
      clf_rf = Pipeline([('vect', CountVectorizer()),
                          ('tfidf', TfidfTransformer()),
                          ('cal', CalibratedClassifierCV(RandomForestClassifier(n_estimators=500, random_state = 123), cv=2, method="sigmoid"),)
      ])
      clf_rf.fit(train['transcript'], train[g])
      predicted = clf_rf.predict(test['transcript'])
      print(metrics.classification_report(test[g], predicted))
      # Impute
      df.loc[inference.index, g] = clf_rf.predict(inference['transcript'])
  

df.to_csv("data/fb_18_20_imputed.csv", index = False)
